Backend/processor.py
import os
import numpy as np
from PIL import Image

# Assume 'processed' folder exists inside 'uploads'
PROCESSED_FOLDER = os.path.join(os.path.dirname(__file__), 'uploads', 'processed')

def process_image_for_upscale(unique_filename_server):
    """
    Handles file I/O, placeholder processing, and saving the final image.
    
    Args:
        unique_filename_server (str): The unique name of the file saved in 'originals/'.
    
    Returns:
        bool: True on success, False on failure.
    """
    
    # 1. Define file paths
    originals_path = os.path.join(os.path.dirname(__file__), 'uploads', 'originals', unique_filename_server)
    processed_path = os.path.join(PROCESSED_FOLDER, unique_filename_server)
    
    try:
        # 2. Open Image and Prepare for Model (Pillow I/O)
        img = Image.open(originals_path).convert("RGB")
        
        # --- START OF AI MODEL INTEGRATION ZONE ---
        
        # 3. Convert Image to NumPy Array (Format needed by most AI models)
        img_array = np.array(img) 
        
        # 4. *[CUSTOM MODEL PROCESSING GOES HERE]*
        # 
        #    Example: Call your custom AI model function:
        #    upscaled_array = your_custom_model.predict(img_array)
        
        # 5. Placeholder: Simulating Upscaling/Processing
        #    For this task, we will just resize it to 2x (simulating output)
        new_width = img.width * 2
        new_height = img.height * 2
        processed_img = img.resize((new_width, new_height), Image.Resampling.LANCZOS)
        
        # --- END OF AI MODEL INTEGRATION ZONE ---

        # 6. Save the Processed Image
        processed_img.save(processed_path)
        
        return True

    except FileNotFoundError:
        print(f"Error: Original file not found at {originals_path}")
        return False
    except Exception as e:
        print(f"Error during image processing: {e}")
        return False

if not os.path.exists(PROCESSED_FOLDER):
    os.makedirs(PROCESSED_FOLDER)