Backend/requirements.txt
NumPy
Pandas
Flask
Flask-CORS
Flask-SQLAlchemy
Pillow